2_Deployment/requirements.txt
https://github.com/explosion/spacy-models/releases/download/fr_core_news_sm-2.3.0/fr_core_news_sm-2.3.0.tar.gz
blis==0.7.4
catalogue==1.0.0
Keras==2.4.3
tensorflow-cpu
Keras-Preprocessing==1.1.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cymem==2.0.5
Flask==1.1.2
scikit-learn==0.23.2
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==1.0.0
MarkupSafe==1.1.1
murmurhash==1.0.5
numpy==1.19.5
pandas
plac==1.1.3
preshed==3.0.5
python-dateutil==2.8.1
pytz==2020.5
requests==2.25.1
scipy
six==1.15.0
spacy==2.3.5
srsly==1.0.5
thinc==7.4.5
threadpoolctl==2.1.0
tqdm==4.56.0
urllib3==1.26.3
wasabi==0.8.1
Werkzeug==1.0.1
